refactor[main]: Modularize server setup and routes
- Splits server setup into modular route handlers for better organization
- Moves environment variable handling to a dedicated config file
- Refactors weather API function to reduce redundancy
- Streamlines WebSocket connections with separated media stream handler
