+++ src/SettingsConfiguration.js
